initialize core editor systems in router (#6967)
* initialize core admin systems in top level of module

* editor initialization
